# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于Gin框架和Gorm二次封装的Go语言Web开发框架",
        "title": "inis CMS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "兔子 or racns",
            "url": "https://github.com/racns/inis",
            "email": "97783391@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "paths": {
...
        "/api/test/{method}": {
            "get": {
                "description": "根据不同方法获取测试相关数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "获取测试数据",
                "parameters": [
                    {
                        "enum": [
                            "request",
                            "alipay"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询条件",
                        "name": "where",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "或条件",
                        "name": "or",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊查询",
                        "name": "like",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "是否使用缓存",
                        "name": "cache",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "根据不同方法更新测试相关数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "更新测试数据",
                "parameters": [
                    {
                        "enum": [
                            "request"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询条件",
                        "name": "where",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "或条件",
                        "name": "or",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊查询",
                        "name": "like",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "根据不同方法提交测试相关数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "提交测试数据",
                "parameters": [
                    {
                        "enum": [
                            "return-url",
                            "notify-url",
                            "request",
                            "upload"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询条件",
                        "name": "where",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "或条件",
                        "name": "or",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊查询",
                        "name": "like",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "根据不同方法删除测试相关数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "删除测试数据",
                "parameters": [
                    {
                        "enum": [
                            "request"
                        ],
                        "type": "string",
                        "description": "方法名",
                        "name": "method",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询条件",
                        "name": "where",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "或条件",
                        "name": "or",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊查询",
                        "name": "like",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "方法调用错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}